Refactor tests and re-add dotenv as a dependency
