+++ pyproject.toml
@@ -14,6 +14,7 @@ panos-upgrade-assurance = "^0.3.1"
 pydantic = "^2.5.3"
 typer = { extras = ["all"], version = "^0.9.0" }
 setuptools = "^69.0.3"
+load-dotenv = "^0.1.0"
 
 [tool.poetry.group.dev.dependencies]
 black = "^23.12.1"
